backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Zoltan Kohalmy <zoltankohalmy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
djangorestframework = "^3.14.0"
pytest-dotenv = "^0.5.2"
pillow = "^10.2.0"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.7.0"
django-stubs = "^4.2.7"
python-dotenv = "^1.0.0"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
